# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SDR CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List user workspaces
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceBody"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found

  # ─── Leads ────────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/leads/{leadId}/messages:
    get:
      operationId: listLeadMessages
      tags: [messages]
      summary: List AI-generated messages for a lead
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: leadId
          in: path
          required: true
          schema:
            type: string
        - name: campaignId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedMessage"
  /workspaces/{workspaceId}/leads/{leadId}/generate-messages:
    post:
      operationId: generateMessages
      tags: [messages]
      summary: Generate AI messages for a lead with a campaign
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMessagesBody"
      responses:
        "200":
          description: Generated messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedMessage"
  /workspaces/{workspaceId}/leads/{leadId}/send-message:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Mark a message as sent (moves lead to Tentando Contato)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: Message sent and lead stage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  # ─── Funnel Stages ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - ownerId
        - createdAt
    CreateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Lead:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        stage:
          type: string
        assigneeId:
          type: ["string", "null"]
        customFieldValues:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - name
        - stage
        - createdAt
        - updatedAt
...
    GeneratedMessage:
      type: object
      properties:
        id:
          type: string
        leadId:
          type: string
        campaignId:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [generated, sent]
        sentAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - leadId
        - campaignId
        - content
        - status
        - createdAt
    GenerateMessagesBody:
      type: object
      properties:
        campaignId:
          type: string
        count:
          type: integer
          default: 3
      required:
        - campaignId
    SendMessageBody:
      type: object
      properties:
        messageId:
          type: string
      required:
        - messageId
...